Replace hearts with healthbar. Player loses health continuously as long (s)he is colliding with the monster.  Furthermore, potion heals all hp.

+++ core/src/com/mygdx/amaze/entities/Player.java
@@ -20,7 +20,7 @@ public class Player {
     public float x;
     public float y;
 
-    public int health = 99;
+    public float health = 99;
     public boolean shielded = false;
 
     //inventory
